macOS 10.15.4 beta 5, Xcode 11.4 beta 3 (iOS, tvOS, watchOS, iOSSupport 13.4 beta)

+++ iOS/Frameworks/PhotosUI.framework/UIScrollView-PhotosUI.h
@@ -10,7 +10,7 @@
 - (void)_pu_endPageSwipeTest:(id)arg1;
 - (void)_pu_incrementForPageSwipeTest:(id)arg1 dt:(double)arg2 stop:(_Bool *)arg3;
 - (void)_pu_beginPageSwipeTest:(id)arg1;
-- (void)pu_ppt_performPageSwipeTest:(id)arg1 iterations:(long long)arg2 numberOfPages:(long long)arg3 scrollAxis:(int)arg4;
+- (void)pu_ppt_performPageSwipeTest:(id)arg1 iterations:(long long)arg2 numberOfPages:(long long)arg3 scrollAxis:(unsigned long long)arg4;
 - (void)pu_ppt_scrollToContentOffset:(struct CGPoint)arg1 withCompletionHandler:(CDUnknownBlockType)arg2;
 - (void)_pu_setPPTScrollHelper:(id)arg1;
 - (id)_pu_pptScrollHelper;
@@ -22,7 +22,7 @@
 - (void)pu_scrollToEdge:(unsigned int)arg1 animated:(_Bool)arg2;
 - (_Bool)_pu_isScrolledAtEdge:(unsigned int)arg1 tolerance:(double)arg2;
 @property(readonly, nonatomic) _Bool pu_isBouncing;
-- (int)_pu_scrollableAxis;
+- (unsigned long long)_pu_scrollableAxis;
 - (_Bool)pu_isScrolledBeyondEdge:(unsigned int)arg1;
 - (_Bool)pu_isScrolledAtEdge:(unsigned int)arg1 tolerance:(double)arg2;
 - (_Bool)pu_isScrolledAtEdge:(unsigned int)arg1;
