+++ macOS/iOSSupport/PrivateFrameworks/UIKitCore.framework/_UICollectionViewLayoutOrthogonalScrolling-Protocol.h
@@ -19,7 +19,7 @@
 - (struct CGRect)_orthogonalScrollingContentRectForSection:(long long)arg1;
 - (struct CGRect)_orthogonalScrollingLayoutRectForSection:(long long)arg1;
 - (double)_orthogonalScrollingDecelerationRateForSection:(long long)arg1;
-- (int)_orthogonalScrollingAxis;
+- (unsigned long long)_orthogonalScrollingAxis;
 - (struct CGPoint)_orthogonalScrollingTargetContentOffsetForOffset:(struct CGPoint)arg1 section:(long long)arg2;
 - (BOOL)_orthogonalScrollingUsesTargetContentOffsetForSection:(long long)arg1;
 - (BOOL)_orthogonalScrollingShouldCenterCustomPagingSizeForSection:(long long)arg1;
